% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subfunctions.R
\name{cal_average_pseudo_atts}
\alias{cal_average_pseudo_atts}
\title{calculate mean pseudo treatment effects}
\usage{
cal_average_pseudo_atts(x, unitname, gname)
}
\description{
Much like the previous functions, this function computes the average of the 
pseudo treatment effects. It is designed to be applied to the dataframe of 
treatment effects for each treated unit separately (using sapply). The 
resulting averages can then be used to test the credibility of PTA.
}
\keyword{internal}
