% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subfunctions.R
\name{cal_group_crit_vals}
\alias{cal_group_crit_vals}
\title{calculate group-wise critical values}
\usage{
cal_group_crit_vals(x)
}
\description{
It takes as input the residuals data for one treated unit (provided to it via
lappy(split())). This data includes all bootstraped residuals for every point
in time. Based on that, the function calculates the critical values (double-sided
left- and right-sided) and the uniform critical value on the group-level. To 
do this, it first takes the average for all bootstraped units. Then it uses 
the 'bstrp' averages to infer percentiles for the confidence intervals. 

For the uniform confidence interval, it extracts the maximum of the 'bstrap' 
averages. Later, these maxima (one for each treated unit/group) provide the 
basis for infering the 95% uniform confidence interval.
}
\keyword{internal}
