% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subfunctions.R
\name{cal_group_time_effects}
\alias{cal_group_time_effects}
\title{calculate group-time average effects}
\usage{
cal_group_time_effects(x, tname)
}
\description{
It takes as input the residuals data for one treated unit (provided to it via
lappy(split())). This data includes all bootstraped residuals for every point
in time. Based on that, the function calculates the critical values (double-sided
left- and right-sided) and the uniform critical value.
}
\keyword{internal}
