% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subfunctions.R
\name{cal_pretreatment_pvalues}
\alias{cal_pretreatment_pvalues}
\title{calculate pretreatment tests}
\usage{
cal_pretreatment_pvalues(x)
}
\description{
This function calculates the critical values of the pseudo treatment effects
and derives a p-value, indicating whether the pseudo pretreatment effects are
different from the pseudo treatment effects of the controls. It is designed
to be applied to the bootstrapped residuals grouped by the treated unit they
were bootstrapped for.
}
\keyword{internal}
