% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subfunctions.R
\name{cal_simple_averages_residuals}
\alias{cal_simple_averages_residuals}
\title{calculate simple average crit values}
\usage{
cal_simple_averages_residuals(x)
}
\description{
This function is designed to be applied to the resized residuals of each 
bootstrap iteration. For each of these iterations, the average resized residual
is taken. Each of these means stands for a hypothetical simple average treatment
effect and can therefore be used to compute critical values later.
}
\keyword{internal}
