% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subfunctions.R
\name{define_control_group}
\alias{define_control_group}
\title{define control group and treated unit}
\usage{
define_control_group(data, gname, tname, group, control_group)
}
\description{
define control group and treated unit
}
\keyword{internal}
