% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subfunctions.R
\name{diff_to_ref}
\alias{diff_to_ref}
\title{Difference to reference value calculator}
\usage{
diff_to_ref(data, tname, yname, pret)
}
\description{
Difference to reference value calculator
}
\keyword{internal}
