% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subfunctions.R
\name{difference_builder}
\alias{difference_builder}
\title{difference_builder}
\usage{
difference_builder(data, idname, tname, yname, pret, universal_base)
}
\description{
difference_builder
}
\keyword{internal}
