% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subfunctions.R
\name{estimate_outcome_model}
\alias{estimate_outcome_model}
\title{outcome model estimator}
\usage{
estimate_outcome_model(
  data,
  tname,
  idname,
  unitname,
  t,
  C,
  G,
  g,
  pret,
  xformula,
  form
)
}
\description{
This function estimates the outcome model and the ATTs for one time period. It is supposed to 
in a loop over all time periods with t being the period of the current iteration. 
It takes the data as an input and returns a list with outcome_residuals and the treatment effects.
}
\keyword{internal}
