% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_staggered_did.R
\name{simple_staggered_did}
\alias{simple_staggered_did}
\title{Staggered DiD with single treated units}
\usage{
simple_staggered_did(
  yname,
  tname,
  gname,
  idname,
  unitname = idname,
  xformula = NA,
  varformula = NA,
  universal_base = FALSE,
  control_group = "never_treated",
  bstrap = 200,
  data
)
}
\arguments{
\item{yname}{Name of the outcome variable. The outcome variable needs to be numeric.}

\item{tname}{Name of the time variable. The time variable needs to be numeric.}

\item{gname}{Name of the variable indicating treatment timing. Needs to be a 
numeric variable indicating the period of first treatment for
 all observations of the treated units and is zero for all untreated units.}

\item{idname}{Name of the variable that identifies the lowest unit level. 
The combination of idname and tname must uniquely identify all observations.}

\item{unitname}{Name of the variable that indicates the desired unit-level of the analysis, 
which should be the treatment level. Treatment effects are calculated at this unit x time level. 
Aggregated treatment effects are aggregated at this unit level. 
Unitname only needs to be specified, when the desired level of aggregation is not the idname. 
When not specified, unitname is assumed to be the same as idname.}

\item{xformula}{A string vector containing all the variables that 
should enter the outcome model as control variables.}

\item{varformula}{A string vector containing all variables that determine 
the heteroscedasticity of the outcome model errors.}

\item{universal_base}{Requires a logical input. If set to TRUE, pretreatment 
periods are evaluated against the first treatment period. If set to FALSE, 
pretreatment periods are evaluated against the following period. 
Default is FALSE. Treatment periods are always evaluated against the 
last pretreatment period.}

\item{control_group}{A string either containing "never_treated" or 
"not_yet_treated". In case of the former, only never-treated units 
are used as controls. In case of the latter, the pretreatment periods 
of not-yet-treated units are added to the pool of controls.}

\item{bstrap}{A number indicating how many control units should be drawn for the bootstrap sample.
When not specified, bstrap defaults to 200.}

\item{data}{A data frame or tibble containing all relevant variables.}
}
\value{
A list containing three data frames. 
The data frames contain the unit x time treatment effects, the unit-level 
average treatment effects, the overall average treatment effect. The data frame
with the unit-level average treatment effects also contains the p-value 
indicating whether the average pseudo treatment effect is significantly different
from zero.
}
\description{
Estimates unit x time treatment effects, averages them on the unit-level and 
over all units. Also provides confidence intervals based on bootstrapped 
treatment effects.
}
\details{
This functions implements a simple difference in difference estimator for 
staggered treatment adoption with few treated units and many controls. 
Following the the recent literature on staggered treatment adoption, this function 
estimates the unit x time treatment effects and aggregates them to derive the
average treatment effects (Callaway & Sant'Anna, 2021). For inference, this function 
implements the bootstrap method proposed by Alvarez & Ferman (2023), which 
produces confidence intervals of the correct size in situations with a 
fixed number of treated units and a large number of control units.
}
\examples{
# generate example data with heteroscedasticity and one control variable
dta <- data.frame(id = rep(1:50, each=10),
                  t = rep(1:10, n=50),
                  z = abs(rep(rnorm(50, mean=0, sd=2), each=10)),
                  x = rnorm(500, mean=4, sd=1))

dta$e <- rnorm(500, mean=0, sd=sqrt(1+dta$z))
dta$y <- 1+ dta$x + dta$e
dta$group <- 0
dta$group[dta$id==1] <- 6
dta$group[dta$id==2] <- 5
dta$y[dta$group <= dta$t] <- dta$y[dta$group <= dta$t] + 4

# run simple_staggered_did
output <- simple_staggered_did(yname = "y",
                               tname = "t",
                               idname = "id",
                               gname = "group",
                               xformula = "x",
                               varformula = "z",
                               data = dta)
}
\references{
Callaway, Brantly and Pedro H.C. Sant'Anna. 2021. "Difference-in-Differences with Multiple Time Periods." Journal of Econometrics, Vol. 225, No. 2, pp. 200-230. <https://arxiv.org/abs/1803.09015>

Alvarez, Luis and Bruno Ferman. 2023. "Extensions for Inference in Difference-in-Difference with Few Treated Clusters". <https://arxiv.org/pdf/2302.03131.pdf>
}
